pyproject.toml
[tool.poetry]
name = "smartfit-v1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.2"}
async-sqlalchemy = "^1.0.0"
dishka = "^1.3.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
alembic = "^1.13.2"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
